amplify/backend/api/pinecontainer/src/requirements.txt
python-dotenv
pinecone-client
langchain
openai
transformers
pypdf
tiktoken
fastapi
typing
pydantic
uvicorn
boto3
